feat(core): BOM-line quantity Integer -> Numeric(10,4)
Structural BOM-line quantities (BOMItem.quantity in bom.py/bom_items_master
-- the canonical structural qty --, BomItem.quantity in bom_item.py,
BomVariantItem.quantity, RevisionBomSnapshotItem.quantity) were Integer,
so a line calling for a fractional unit of measure (2.5 meters of wire,
0.5 kg of adhesive) could not be represented at all -- the DBAPI rejects
binding a Decimal to an Integer column outright. Switches all four to
Numeric(10,4), matching the precision already used for
quantity_on_hand/quantity_required/quantity_issued/quantity_returned in
inventory.py/work_order.py and for quantity in mbom.py/service_bom.py/part.py.

Genuine whole-unit counts (work_order.quantity_ordered/completed/scrapped,
kanban.reorderQuantity, contract pricing-tier breakpoints, deviation/ECO
affected-unit counts) are left as Integer -- they are not structural BOM/
consumption quantities.

Fixes a regression this surfaced in bom_service.py's effective-quantity
rollup (the P0 BOM-explosion fix): _compute_levels_and_effective_qty
multiplied line quantities down the assembly tree, and get_cost_rollup then
multiplied that by a float unit cost -- float * Decimal raises TypeError
once quantity is Decimal. Effective quantity is now cast to float once at
the leaf, consistent with the float-based cost arithmetic throughout that
module.

Adds chained alembic migration 034 (single head confirmed).

TDD: app/tests/test_qty_precision.py creates a BOM line (and template/
variant/revision-snapshot line) with quantity=2.5, round-trips it through a
fresh session, and asserts it comes back as Decimal(2.5000) -- previously
failed with sqlite3.ProgrammingError: type decimal.Decimal is not
supported (Integer column cannot bind a Decimal at all).

+++ backend/app/models/bom.py
@@ -56,7 +56,7 @@ class BOMItem(Base, TenantAwareMixin):
     id = Column(Integer, primary_key=True)
     bom_id = Column(Integer, ForeignKey("boms.id", ondelete="CASCADE"), nullable=False, index=True)
     part_id = Column(Integer, ForeignKey("parts.id", ondelete="CASCADE"), index=True)
-    quantity = Column(Integer, default=1)
+    quantity = Column(Numeric(10, 4), default=1)
     unit = Column(String, default="EA")
     reference_designator = Column(String)
     sort_order = Column(Integer, default=0)
